Gemini ChatBot/requirements.txt
google-generativeai
python-dotenv
ipython==8.16.1
streamlit
ipykernel
pymongo